requirements.txt
Flask
flask-cors
python-dotenv
flask-sqlalchemy
marshmallow
gunicorn
